+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Unreleased
+------------
+* Upgrading to docker-based deployment
+* Adding CustomDotEnvParameterArn parameter
+* Adding CustomConfigJsParameterArn parameter
+* Adding CustomInterfaceConfigJsParameterArn parameter
 
 3.0.0
 -----